This is quite a significant update to the experimental eclass.  I'd also like to propose a policy change which this update enables.
To now, the multilib source dir env vars have only been set during the
initialisation of the build dir for each ABI, and have been local to
that pass.  They are now set globally (even when multilib is disabled),
and are available for export to ebuilds which can make use of them to
find their documentation and configuration files during the src_install
phase when those files are not copied to the build dir.  This will
hopefully eventually allow the removal of the documentaion copying hack.

This update also renames those variables to more meaningful names in
this context, along with a few other renames to various functions.