Update bin/add_multilib_abi to fix the environment.bz2 files in /var/db/pkg instead of just the db keys.
When editing the USE and IUSE db keys but not changing the variables in environment.bz2, packages installed with an old portage-multilib installation, the eclass implementation, or by vanilla portage the environment file would be inconsistent with what portage sees. This means that when multilib-portage tries to uninstall a package, the pkg_postrm() stage would always die() with the message ``Could not determine or profile ABIs''. This new script updates the bzip2-ed environment file's USE and IUSE environment variables properly and fixes this problem.

http://forums.gentoo.org/viewtopic-p-6710761.html#6710761

Signed-off-by: Thomas Sachau <tommy@gentoo.org>